# backend/src/modules/projects/types/dtos/requests/create-project-req.dto.ts
import { ApiProperty } from "@nestjs/swagger"
import { IsNotEmpty, IsNumber, IsString } from "class-validator"

export class CreateProjectReqDto {
    @ApiProperty()
    @IsString()
    @IsNotEmpty()
    class: string

    @ApiProperty()
    @IsNumber()
    @IsNotEmpty()
    currentYear: number

    @ApiProperty()
    @IsString()
    @IsNotEmpty()
    topic: string

    @ApiProperty()
    @IsString()
    @IsNotEmpty()
    ppiId: string
}