# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/subjects": {
      "post": {
        "operationId": "SubjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectWithTeacherResDto"
                }
              }
            }
          }
        },
        "tags": [
          "subjects"
        ]
      },
      "get": {
        "operationId": "SubjectController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResDto"
                }
              }
            }
          }
        },
        "tags": [
          "subjects"
        ]
      }
    },
    "/subjects/{id}": {
      "get": {
        "operationId": "SubjectController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectWithTeacherResDto"
                }
              }
            }
          }
        },
        "tags": [
          "subjects"
        ]
      },
      "put": {
        "operationId": "SubjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectWithTeacherResDto"
                }
              }
            }
          }
        },
        "tags": [
          "subjects"
        ]
      },
      "delete": {
        "operationId": "SubjectController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "subjects"
        ]
      }
    },
...
    "/projects": {
      "post": {
        "operationId": "ProjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResDto"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "ProjectController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "ppiId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NOT_STARTED",
                "STARTED",
                "FINISHED"
              ]
            }
          },
          {
            "name": "currentYear",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "class",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topic",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResDto"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectFullResDto"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ]
      },
      "put": {
        "operationId": "ProjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResDto"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "operationId": "ProjectController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/content/{id}": {
      "patch": {
        "operationId": "ProjectController_updateContentByID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectContentReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectFullResDto"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/status/{id}": {
      "patch": {
        "operationId": "ProjectController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResDto"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ]
      }
    }
  },
  "info": {
    "title": "SIPPIE",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MetadataDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "itemsPerPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "itemsPerPage",
          "totalPages",
          "totalItems"
        ]
      },
      "PaginationResDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataDto"
          }
        },
        "required": [
          "items",
          "metadata"
        ]
      },
...
      "randomUUID": {
        "type": "object",
        "properties": {}
      },
...
      "CreateSubjectReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "teachers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "teachers"
        ]
      },
      "TeacherResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "SubjectWithTeacherResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "teachers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeacherResDto"
            }
          },
          "courseId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "teachers",
          "courseId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateSubjectReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "teachers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "teachers"
        ]
      },
...
      "CreateProjectReqDto": {
        "type": "object",
        "properties": {
          "class": {
            "type": "string"
          },
          "currentYear": {
            "type": "number"
          },
          "topic": {
            "type": "string"
          },
          "ppiId": {
            "type": "string"
          }
        },
        "required": [
          "class",
          "currentYear",
          "topic",
          "ppiId"
        ]
      },
      "ProjectResDto": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/randomUUID"
          },
          "class": {
            "type": "string"
          },
          "currentYear": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "STARTED",
              "FINISHED"
            ]
          },
          "topic": {
            "type": "string"
          },
          "ppiId": {
            "$ref": "#/components/schemas/randomUUID"
          },
          "ppiClassPeriod": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "class",
          "currentYear",
          "status",
          "topic",
          "ppiId",
          "ppiClassPeriod",
          "createdBy",
          "updatedBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProjectFullResDto": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/randomUUID"
          },
          "topic": {
            "type": "string"
          },
          "scope": {
            "type": "object"
          },
          "justification": {
            "type": "object"
          },
          "generalObjective": {
            "type": "object"
          },
          "specificObjectives": {
            "type": "object"
          },
          "subjectsContributions": {
            "type": "object"
          },
          "methodology": {
            "type": "object"
          },
          "timeline": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "topic"
        ]
      },
      "UpdateProjectReqDto": {
        "type": "object",
        "properties": {
          "class": {
            "type": "string"
          },
          "currentYear": {
            "type": "number"
          },
          "topic": {
            "type": "string"
          }
        },
        "required": [
          "class",
          "currentYear",
          "topic"
        ]
      },
      "UpdateProjectContentReqDto": {
        "type": "object",
        "properties": {
          "topic": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "justification": {
            "type": "string"
          },
          "generalObjective": {
            "type": "string"
          },
          "specificObjectives": {
            "type": "string"
          },
          "subjectsContributions": {
            "type": "string"
          },
          "methodology": {
            "type": "string"
          },
          "timeline": {
            "type": "string"
          }
        },
        "required": [
          "topic"
        ]
      },
      "ChangeStatusReqDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "STARTED",
              "FINISHED"
            ]
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}